By default body should be quoted-printable not base64
Making the body use quoted printable by default allows for easier
debugging by looking at the message as it traverses the mail server, and
there is no reason why the body should be treated like an attachment and
base64 encoded.

A new test has been introduced as well that actually verifies that the
body has not been base64 encoded. The test passes when quoted-printable
is set, so hopefully we can avoid future regressions.