Set a default message sender if no sender set in DebugMailer
Use-case: I construct Messages in my code without a sender, because I want to
use the ``default_sender`` as the sender. Works as a charm with ``Mailer``.

However, when I use ``DebugMailer`` with these Messages, I get a validation
error, because the sender is not set. The ``Mailer`` class takes care of
setting the sender if default sender is set, but the ``DebugMailer`` does not
do that. This patch then sets a dummy sender, if no sender is set in Message,
to allow using ``DebugMailer`` with sender-less Messages.